Сделан расчёт медицинских рисков и также добавлен расчёт нескольких рисков сразу и тесты к ним

+++ src/main/java/org/khamit/travel/insurance/core/repository/InsuranceLimitRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.Query;
 public interface InsuranceLimitRepository extends JpaRepository<InsuranceLimit, Long> {
 
     @Query("select a from InsuranceLimit a where a.limitMin<=?1 and a.limitMax>=?1")
-    InsuranceLimit findCoefByLimit(int limit);
+    InsuranceLimit findCoefByLimit(Double limit);
 }
